% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureReg_methods.R
\name{logLik.mixtureReg}
\alias{logLik.mixtureReg}
\title{Obtain Log-likelihood from a mixtureReg Object}
\usage{
\method{logLik}{mixtureReg}(mixtureModel)
}
\arguments{
\item{mixtureModel}{mixtureReg object, typically result from 'mixtureReg()'.}
}
\value{
Return a numeric value of log likelihood.
}
\description{
S3 method for class 'mixtureReg'.
However, it doesn't return a 'logLik' object.
For simlicity, it returns a 'numeric' value.
}

