% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureReg_methods.R
\name{orderedLines}
\alias{orderedLines}
\title{Sort by X Coordinates and Add Line to a Plot}
\usage{
orderedLines(x, y, ...)
}
\arguments{
\item{x}{X coordinate vectors of points to join.}

\item{y}{Y coordinate vectors of points to join.}

\item{...}{Further graphical parameters.}
}
\description{
Rearrange X and Y coordinates before calling "lines()" function.
}

