% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureReg_methods.R
\name{plot.mixtureReg}
\alias{plot.mixtureReg}
\title{Plot Fit and Mixing Probability of a mixtureReg Object}
\usage{
\method{plot}{mixtureReg}(mixtureModel, which = 1:2, xName = NULL,
  yName = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{mixtureModel}{mixtureReg object, typically result from 'mixtureReg()'.}

\item{which}{numeric; choose which plot to display. '1' gives a plot of fit; '2' gives a plot of mixing probability.}

\item{xName}{character; Name used to pick x variable from data.}

\item{yName}{character; Name used to pick y variable from data.}

\item{xlab}{character; label that should be put on the x axis.}

\item{ylab}{character; label that should be put on the y axis.}

\item{...}{Further graphical parameters.}
}
\description{
S3 plot method for class 'mixtureReg'.
}

