% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureReg.R
\name{mixtureReg}
\alias{mixtureReg}
\title{Function to Fit Mixture of Regressions}
\usage{
mixtureReg(regData, formulaList, xName = NULL, yName = NULL,
  mixingProb = c("constant", "kern"), initialWList = NULL,
  epsilon = 1e-08, max_iter = 10000, max_restart = 15,
  min_lambda = 0.01, min_sigmaRatio = 0.1, silently = FALSE)
}
\arguments{
\item{regData:}{data frame used in fitting model.}

\item{formulaList:}{a list of the regression components that need to be estimated.}

\item{initialWList:}{a list of weights guesses (provided by user).}
}
\value{
A class 'mixtureReg' object.
}
\description{
The main function in this package.
}

